Remove unused CartItem and OrderItem controllers and restructure database
- reorganized the data model by removing the CartItemController, OrderItemController, and their corresponding service classes.
- simplified relationships in the database and corrected mappings in entity classes for cleaner design and enhanced clarity.
