--- src/main/java/com/achdev/onlinebookstoreapp/controller/CartItemController.java
@@ -1,76 +1,0 @@
-package com.achdev.onlinebookstoreapp.controller;
-
-import com.achdev.onlinebookstoreapp.dto.cart.item.UpdateCartItemRequestDto;
-import com.achdev.onlinebookstoreapp.dto.errors.CommonApiErrorResponse;
-import com.achdev.onlinebookstoreapp.service.CartItemService;
-import com.achdev.onlinebookstoreapp.util.ApiResponseConstants;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-@Tag(name = "Cart item management", description = "Endpoints for managing cart items")
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/cart/cart-items")
-public class CartItemController {
-    private final CartItemService cartItemService;
-
-    @Operation(
-            summary = ApiResponseConstants.CART_ITEM_UPDATE_DESCRIPTION,
-            description = ApiResponseConstants.CART_ITEM_UPDATE_DESCRIPTION,
-            responses = {
-                    @ApiResponse(
-                            responseCode = ApiResponseConstants.RESPONSE_CODE_OK,
-                            description = "Successfully updated cart item's information"
-                    ),
-                    @ApiResponse(responseCode = ApiResponseConstants.RESPONSE_CODE_NOT_FOUND,
-                            description = ApiResponseConstants.NOT_FOUND_DESCRIPTION,
-                            content = @Content(schema = @Schema(implementation =
-                                    CommonApiErrorResponse.class))
-                    ),
-                    @ApiResponse(responseCode = ApiResponseConstants.RESPONSE_CODE_FORBIDDEN,
-                            description = ApiResponseConstants.FORBIDDEN_DESCRIPTION,
-                            content = @Content(schema = @Schema(
-                                    implementation = CommonApiErrorResponse.class))
-                    )
-            }
-    )
-    @PreAuthorize("hasRole('ROLE_USER')")
-    @PutMapping("/{cartItemId}")
-    public void updateCartItem(
-            @PathVariable Long cartItemId,
-            @RequestBody @Valid UpdateCartItemRequestDto requestDto) {
-        cartItemService.updateCartItem(cartItemId, requestDto);
-    }
-
-    @Operation(
-            summary = ApiResponseConstants.CART_ITEM_DELETE_DESCRIPTION,
-            description = ApiResponseConstants.CART_ITEM_DELETE_DESCRIPTION,
-            responses = {
-                    @ApiResponse(responseCode = ApiResponseConstants.RESPONSE_CODE_FORBIDDEN,
-                            description = ApiResponseConstants.FORBIDDEN_DESCRIPTION,
-                            content = @Content(schema = @Schema(
-                                    implementation = CommonApiErrorResponse.class))
-                    )
-            }
-    )
-    @PreAuthorize("hasRole('ROLE_USER')")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    @DeleteMapping("/{cartItemId}")
-    public void deleteCartItem(@PathVariable Long cartItemId) {
-        cartItemService.deleteCartItem(cartItemId);
-    }
-}
