+++ src/main/java/com/achdev/onlinebookstoreapp/service/ShoppingCartService.java
@@ -1,6 +1,7 @@
 package com.achdev.onlinebookstoreapp.service;
 
 import com.achdev.onlinebookstoreapp.dto.cart.item.CartItemRequestDto;
+import com.achdev.onlinebookstoreapp.dto.cart.item.UpdateCartItemRequestDto;
 import com.achdev.onlinebookstoreapp.dto.shopping.cart.ShoppingCartDto;
 import com.achdev.onlinebookstoreapp.model.ShoppingCart;
 import com.achdev.onlinebookstoreapp.model.User;
@@ -15,4 +16,8 @@ public interface ShoppingCartService {
     ShoppingCart getShoppingCartByUserEmail(String email);
 
     void clearShoppingCart(ShoppingCart shoppingCart);
+
+    void updateCartItem(Long cartItemId, UpdateCartItemRequestDto requestDto);
+
+    void deleteCartItem(Long cartItemId);
 }
