+++ src/main/resources/db/changelog/changes/11-create-order-table.yaml
@@ -1,10 +1,10 @@
 databaseChangeLog:
   - changeSet:
-      id: create-order_items-table
+      id: create-order-table
       author: andrey chertykovtsev
       changes:
         - createTable:
-            tableName: order_items
+            tableName: orders
             columns:
               - column:
                   name: id
@@ -14,23 +14,28 @@ databaseChangeLog:
                     primaryKey: true
                     nullable: false
               - column:
-                  name: orders_id
+                  name: user_id
                   type: BIGINT
                   constraints:
                     nullable: false
               - column:
-                  name: book_id
-                  type: BIGINT
+                  name: status
+                  type: VARCHAR(255)
                   constraints:
                     nullable: false
               - column:
-                  name: quantity
-                  type: INT
+                  name: total
+                  type: DECIMAL
                   constraints:
                     nullable: false
               - column:
-                  name: price
-                  type: DECIMAL(10, 2)
+                  name: order_date
+                  type: DATETIME
+                  constraints:
+                    nullable: false
+              - column:
+                  name: shipping_address
+                  type: VARCHAR(255)
                   constraints:
                     nullable: false
               - column:
@@ -40,14 +45,8 @@ databaseChangeLog:
                   constraints:
                     nullable: false
         - addForeignKeyConstraint:
-            baseColumnNames: orders_id
-            baseTableName: order_items
-            constraintName: FK_ORDER_ITEMS_ON_ORDERS
+            baseColumnNames: user_id
+            baseTableName: orders
+            constraintName: fk_orders_on_user
             referencedColumnNames: id
-            referencedTableName: orders
-        - addForeignKeyConstraint:
-            baseColumnNames: book_id
-            baseTableName: order_items
-            constraintName: FK_ORDER_ITEMS_ON_BOOKS
-            referencedColumnNames: id
-            referencedTableName: books
+            referencedTableName: users
