--- src/main/resources/db/changelog/changes/11-create-shopping_carts_cart_items.yaml
@@ -1,34 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: create-shopping_carts_cart_items
-      author: andrey chertykovtsev
-      changes:
-        - createTable:
-            tableName: shopping_carts_cart_items
-            columns:
-              - column:
-                  name: shopping_carts_id
-                  type: BIGINT
-                  constraints:
-                    nullable: false
-              - column:
-                  name: cart_items_id
-                  type: BIGINT
-                  constraints:
-                    nullable: false
-        - addUniqueConstraint:
-            columnNames: cart_items_id
-            constraintName: uc_shopping_carts_items_cart_item
-            tableName: shopping_carts_cart_items
-        - addForeignKeyConstraint:
-            constraintName: fc_shopping_carts_cart_items_on_shopping_carts
-            baseColumnNames: shopping_carts_id
-            baseTableName: shopping_carts_cart_items
-            referencedColumnNames: user_id
-            referencedTableName: shopping_carts
-        - addForeignKeyConstraint:
-            constraintName: fc_shopping_carts_cart_items_on_cart_items
-            baseColumnNames: cart_items_id
-            baseTableName: shopping_carts_cart_items
-            referencedColumnNames: id
-            referencedTableName: cart_items
