+++ src/main/resources/db/changelog/changes/12-create-order_items-table.yaml
@@ -1,41 +1,36 @@
 databaseChangeLog:
   - changeSet:
-      id: create-order-table
+      id: create-order_items-table
       author: andrey chertykovtsev
       changes:
         - createTable:
-            tableName: orders
+            tableName: order_items
             columns:
               - column:
-                  name:  id
+                  name: id
                   type: BIGINT
                   autoIncrement: true
                   constraints:
                     primaryKey: true
                     nullable: false
               - column:
-                  name: user_id
+                  name: order_id
                   type: BIGINT
                   constraints:
                     nullable: false
               - column:
-                  name: status
-                  type: VARCHAR(255)
+                  name: book_id
+                  type: BIGINT
                   constraints:
                     nullable: false
               - column:
-                  name: total
-                  type: DECIMAL
+                  name: quantity
+                  type: INT
                   constraints:
                     nullable: false
               - column:
-                  name: order_date
-                  type: DATETIME
-                  constraints:
-                    nullable: false
-              - column:
-                  name: shipping_address
-                  type: VARCHAR(255)
+                  name: price
+                  type: DECIMAL(10, 2)
                   constraints:
                     nullable: false
               - column:
@@ -45,8 +40,14 @@ databaseChangeLog:
                   constraints:
                     nullable: false
         - addForeignKeyConstraint:
-            baseColumnNames: user_id
-            baseTableName: orders
-            constraintName: fk_orders_on_user
+            baseColumnNames: order_id
+            baseTableName: order_items
+            constraintName: FK_ORDER_ITEMS_ON_ORDERS
             referencedColumnNames: id
-            referencedTableName: users
+            referencedTableName: orders
+        - addForeignKeyConstraint:
+            baseColumnNames: book_id
+            baseTableName: order_items
+            constraintName: FK_ORDER_ITEMS_ON_BOOKS
+            referencedColumnNames: id
+            referencedTableName: books
