--- src/main/resources/db/changelog/changes/14-create-orders_order_items-table.yaml
@@ -1,34 +1,0 @@
-databaseChangeLog:
-  - changeSet: 
-      id: create-orders_order_items-table
-      author: andrey chertykovtsev
-      changes:
-        - createTable: 
-            tableName: orders_order_items
-            columns:
-              - column: 
-                  name: orders_id
-                  type: BIGINT
-                  constraints:
-                    nullable: false
-              - column: 
-                  name: order_items_id
-                  type: BIGINT
-                  constraints:
-                    nullable: false
-        - addUniqueConstraint:
-            columnNames: order_items_id
-            constraintName: UC_ORDERS_ITEMS_ORDER_ITEMS
-            tableName: orders_order_items
-        - addForeignKeyConstraint:
-            constraintName: fc_orders_order_items_on_orders
-            baseColumnNames: orders_id
-            baseTableName: orders_order_items
-            referencedColumnNames: id
-            referencedTableName: orders
-        - addForeignKeyConstraint:
-            constraintName: fc_orders_order_items_on_order_items
-            baseColumnNames: order_items_id
-            baseTableName: orders_order_items
-            referencedColumnNames: id
-            referencedTableName: order_items
